Cleanup code, split into files, define scope
